Change `model_total_valid_times` attrib to int32

+++ core/ioMod.py
@@ -162,7 +162,7 @@ class OutputObj:
                     break
 
                 try:
-                    idOut.model_total_valid_times = float(ConfigOptions.actual_output_steps)
+                    idOut.model_total_valid_times = np.int32(ConfigOptions.actual_output_steps)
                 except:
                     ConfigOptions.errMsg = "Unable to create total_valid_times global attribute in: " + self.outPath
                     err_handler.log_critical(ConfigOptions, MpiConfig)
